msp430/dac12: leftover artifact moved from msp430dac12p to hpl.
a leftover artifact from testing was moved from msp430dac12p init to
a more logical and correct place in hplmsp430dac12p.

originally, init turned on the adc reference, which isn't reasonable
for many applications.  this line was moved to the hpl-level
configuration for the reference source; if the internal adc vref+ is
selected there, then the adc reference is turned on, and the level is
set to 2.5v.